Add #1 Two Sum tests
Two sets of tests were added for the #1 problem as there are two ways of
solving this problem that I can think of off the top of my head.

The first solution involves a nested loop brute force method to compare
each of the elements of the provided nums array against itself. The
solution array can be returned as soon as the first pair of numbers that
fulfill the target sum are found.

The second solution involves a the use of a hash map data structure in
order to reduce time complexity of the brute force method. A map can be
used to store each element of the nums array and the difference between
it and the target value. Each subsequent element can be compared against
the map to see if the element exists as a difference. If so, the
solution pair can be returned as soon as the pair is found.